requirements.txt
# conda create --name rag python==3.11
langchain_openai
langchain_community
langchain_core
langchain_text_splitters
streamlit
langserve
fastapi
unicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu
langchain_ollama
arxiv
groq
cassio
langchain_groq
wikipedia
langchainhub
python-dotenv
pydantic
langgraph
pandas
numpy
scikit-learn